// frontend/src/app/api/model/userProfileDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserAccountDto } from './userAccountDto';
import { UserPenaltyDto } from './userPenaltyDto';
import { UserSiteDto } from './userSiteDto';


export interface UserProfileDto { 
    matricule?: string;
    firstName?: string;
    lastName?: string;
    email?: string;
    birthDate?: string;
    level?: string;
    isActive?: boolean;
    roleId?: number;
    account?: UserAccountDto;
    penalties?: Array<UserPenaltyDto>;
    sites?: Array<UserSiteDto>;
}