// frontend/src/app/api/model/userPenaltyDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserPenaltyDto { 
    tr?: number;
    reason?: string;
    startDate?: string;
    endDate?: string;
    isActive?: boolean;
    description?: string;
}